requirements.txt
streamlit
numpy
matplotlib
pandas
db-dtypes
gcsfs
st-files-connection
google-cloud-bigquery
pytz